@@ -1,7 +1,8 @@
 {
   "name": "@lapo/asn1js",
-  "version": "1.3.0",
+  "version": "2.0.1",
   "description": "Generic ASN.1 parser/decoder that can decode any valid ASN.1 DER or BER structures.",
+  "type": "module",
   "main": "asn1.js",
   "repository": {
     "type": "git",
@@ -14,13 +15,13 @@
   "homepage": "https://lapo.it/asn1js/",
   "files": [ "asn1.js", "base64.js", "hex.js", "int10.js", "oids.js" ],
   "scripts": {
-    "lint": "npx eslint asn1.js base64.js hex.js int10.js oids.js tags.js index.js parseRFC.js dumpASN1.js",
+    "lint": "npx eslint asn1.js base64.js hex.js int10.js oids.js tags.js context.js index.js parseRFC.js dumpASN1.js",
     "lint-action": "npx @action-validator/cli .github/workflows/node.js.yml",
     "serve": "echo 'Connect to http://localhost:3000/' ; npx statik --port 3000 .",
     "test": "node test"
   },
   "engines": {
-    "node": ">=6.4.0"
+    "node": ">=12.20.0"
   },
   "devDependencies": {
     "eslint": "^8.34.0"
@@ -28,9 +29,12 @@
   "eslintConfig": {
     "env": {
       "es6": true,
-      "amd": true,
       "browser": true,
       "node": true
+    },
+    "parserOptions": {
+      "ecmaVersion": 2015,
+      "sourceType": "module"
     },
     "extends": [ "eslint:recommended" ],
     "globals": {
@@ -38,7 +42,7 @@
     },
     "rules": {
       "strict": [ "error", "function" ],
-      "indent": [ "error", 4, { "ignoredNodes": [ "Program > ExpressionStatement > CallExpression > FunctionExpression > BlockStatement > ExpressionStatement[directive='use strict']:first-child" ] } ],
+      "indent": [ "error", 4 ],
       "linebreak-style": [ "error", "unix" ],
       "semi": [ "warn", "always" ],
       "quotes": [ "error", "single", { "avoidEscape": true } ],
@@ -71,6 +75,11 @@
           "indent": [ "error", 2, { "ignoredNodes": [ "Program > ExpressionStatement > CallExpression > FunctionExpression > BlockStatement > ExpressionStatement[directive='use strict']:first-child" ] } ],
           "comma-dangle": "off",
           "quotes": [ "warn", "double" ]
+        }
+      }, {
+        "files": [ "defs.js" ],
+        "parserOptions": {
+          "ecmaVersion": 2021
         }
       }
     ]
